separate bin and lib
